git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@2622 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/imagingfcs/qfrdrimagingfcs.cpp
@@ -879,10 +879,10 @@ bool QFRDRImagingFCSPlugin::readEvalSettings(const QString &evalFilename, bool i
                     initParams["SEGMENTS"]=value.toInt();
                     paramsReadonly<<"SEGMENTS";
                 } else if (name=="frametime (ms)") {
-                    initParams["FRAMETIME_MS"]=value.toInt();
+                    initParams["FRAMETIME_MS"]=value.toDouble();
                     paramsReadonly<<"FRAMETIME_MS";
                 } else if (name=="frametime (s)") {
-                    initParams["FRAMETIME"]=value.toInt();
+                    initParams["FRAMETIME"]=value.toDouble();
                     paramsReadonly<<"FRAMETIME";
                 } else if (name=="correlator s") {
                     initParams["CORR_S"]=value.toInt();
